fix delete ticket before delete show, user, ticket type

+++ src/controllers/showController.js
@@ -1,4 +1,4 @@
-const { Show, TicketType } = require("../models/model");
+const { Show, TicketType, Ticket } = require("../models/model");
 
 const showController = {
   addShow: async (req, res) => {
@@ -49,15 +49,25 @@ const showController = {
   deleteShow: async (req, res) => {
     try {
       const show = await Show.findById(req.params.id);
+      const tickets = await Ticket.find({ show: req.params.id });
+      const ticketTypes = await TicketType.find({ show: req.params.id });
       if (!show) {
         res.status(404).send({ success: false, message: "Show not found" });
       } else {
-        const { ticketTypes } = await Show.findById(req.params.id);
-        ticketTypes.forEach(async (ticketType) => {
-          await TicketType.findByIdAndDelete(ticketType);
-        });
-        const deletedShow = await Show.findByIdAndDelete(req.params.id);
-        res.status(200).json(deletedShow);
+        if (tickets.length === 0 && ticketTypes === 0) {
+          const { ticketTypes } = await Show.findById(req.params.id);
+          ticketTypes.forEach(async (ticketType) => {
+            await TicketType.findByIdAndDelete(ticketType);
+          });
+          const deletedShow = await Show.findByIdAndDelete(req.params.id);
+          res.status(200).json(deletedShow);
+        } else {
+          res.status(404).send({
+            success: false,
+            message:
+              "delete tickets, ticket types of this show before delete this show",
+          });
+        }
       }
     } catch (error) {
       res.status(500).json(error);
